+++ src/paymongo.js
@@ -1,6 +1,6 @@
-"use strict";
+'use strict';
 
-const services = require("./services");
+const services = require('./services');
 
 function Paymongo(apiKey) {
   if (!(this instanceof Paymongo)) {
@@ -8,8 +8,8 @@ function Paymongo(apiKey) {
   }
 
   this.apiKey = apiKey;
-  this.baseUrl = "https://api.paymongo.com";
-  this.apiVersion = "v1";
+  this.baseUrl = 'https://api.paymongo.com';
+  this.apiVersion = 'v1';
 
   // TODO: Move this to a factory.
   for (let service in services) {