+++ src/services/Base.js
@@ -1,8 +1,8 @@
-const HttpClient = require("../HttpClient");
+const HttpClient = require('../HttpClient');
 
 function BaseService(client) {
   this.httpClient = new HttpClient({
-    baseUrl: client.baseUrl + "/" + client.apiVersion,
+    baseUrl: client.baseUrl + '/' + client.apiVersion,
     apikey: client.apiKey,
   });
 }