+++ src/services/Customer.js
@@ -1,17 +1,17 @@
-const BaseService = require("./Base");
-const Customer = require("../entities/Customer");
+const BaseService = require('./Base');
+const Customer = require('../entities/Customer');
 
 function CustomerService(client) {
   BaseService.call(this, client);
 
-  this.uri = "/customers";
+  this.uri = '/customers';
 }
 
 CustomerService.prototype.retrieve = function(id) {
   return this.httpClient
     .request({
-      url: this.uri + "/" + id,
-      method: "GET",
+      url: this.uri + '/' + id,
+      method: 'GET',
     })
     .then(function(response) {
       return new Promise(function(resolve, _reject) {
@@ -24,7 +24,7 @@ CustomerService.prototype.create = function(params) {
   return this.httpClient
     .request({
       url: this.uri,
-      method: "post",
+      method: 'POST',
       params,
     })
     .then(function(response) {
@@ -37,8 +37,8 @@ CustomerService.prototype.create = function(params) {
 CustomerService.prototype.update = function(id, params) {
   return this.httpClient
     .request({
-      url: this.uri + "/" + id,
-      method: "put",
+      url: this.uri + '/' + id,
+      method: 'PUT',
       params,
     })
     .then(function(response) {
@@ -51,8 +51,8 @@ CustomerService.prototype.update = function(id, params) {
 CustomerService.prototype.delete = function(id) {
   return this.httpClient
     .request({
-      url: this.uri + "/" + id,
-      method: "delete",
+      url: this.uri + '/' + id,
+      method: 'DELETE',
     })
     .then(function(response) {
       return new Promise(function(resolve, _reject) {