+++ src/services/PaymentIntent.js
@@ -1,17 +1,17 @@
-const BaseService = require("./Base");
-const PaymentIntent = require("../entities/PaymentIntent");
+const BaseService = require('./Base');
+const PaymentIntent = require('../entities/PaymentIntent');
 
 function PaymentIntentService(client) {
   BaseService.call(this, client);
 
-  this.uri = "/payment_intents";
+  this.uri = '/payment_intents';
 }
 
 PaymentIntentService.prototype.create = function(params) {
   return this.httpClient
     .request({
       url: this.uri,
-      method: "post",
+      method: 'POST',
       params,
     })
     .then(function(response) {
@@ -24,8 +24,8 @@ PaymentIntentService.prototype.create = function(params) {
 PaymentIntentService.prototype.retrieve = function(id) {
   return this.httpClient
     .request({
-      url: this.uri + "/" + id,
-      method: "GET",
+      url: this.uri + '/' + id,
+      method: 'GET',
     })
     .then(function(response) {
       return new Promise(function(resolve, _reject) {
@@ -37,8 +37,8 @@ PaymentIntentService.prototype.retrieve = function(id) {
 PaymentIntentService.prototype.capture = function(id, params) {
   return this.httpClient
     .request({
-      url: this.uri + "/" + id + "/capture",
-      method: "post",
+      url: this.uri + '/' + id + '/capture',
+      method: 'POST',
       params,
     })
     .then(function(response) {
@@ -51,8 +51,8 @@ PaymentIntentService.prototype.capture = function(id, params) {
 PaymentIntentService.prototype.cancel = function(id) {
   return this.httpClient
     .request({
-      url: this.uri + "/" + id + "/cancel",
-      method: "post",
+      url: this.uri + '/' + id + '/cancel',
+      method: 'POST',
     })
     .then(function(response) {
       return new Promise(function(resolve, _reject) {
@@ -60,6 +60,23 @@ PaymentIntentService.prototype.cancel = function(id) {
       });
     });
 };
+
+PaymentIntentService.prototype.attach = function(id, params) {
+  return this.httpClient
+    .request({
+      url: this.uri + '/' + id + '/attach',
+      method: 'POST',
+      params : {
+        ...params,
+        origin: 'node'
+      },
+    })
+    .then(function(response) {
+      return new Promise(function(resolve, _reject) {
+        resolve(new PaymentIntent(response));
+      });
+    });
+}
 
 Object.setPrototypeOf(PaymentIntentService.prototype, BaseService.prototype);