+++ src/services/Source.js
@@ -1,17 +1,17 @@
-const BaseService = require("./Base");
-const Source = require("../entities/Source");
+const BaseService = require('./Base');
+const Source = require('../entities/Source');
 
 function SourceService(client) {
   BaseService.call(this, client);
 
-  this.uri = "/sources";
+  this.uri = '/sources';
 }
 
 SourceService.prototype.create = function(params) {
   return this.httpClient
     .request({
       url: this.uri,
-      method: "post",
+      method: 'POST',
       params,
     })
     .then(function(response) {
@@ -24,8 +24,8 @@ SourceService.prototype.create = function(params) {
 SourceService.prototype.retrieve = function(id) {
   return this.httpClient
     .request({
-      url: this.uri + "/" + id,
-      method: "GET",
+      url: this.uri + '/' + id,
+      method: 'GET',
     })
     .then(function(response) {
       return new Promise(function(resolve, _reject) {