+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
-## 0.0.0 - 2022-11-21
+## [10.18.0] - 2023-08-07
+
+- Add origin on payment method attachment for tracking SDK usage.
+- Fix PaymentIntent entity for undefined payments.
+
+## [0.0.0] - 2022-11-21
 
 - Initial alpha release.